requirements.txt
pymongo
pandas
apache-airflow-providers-mongo  
